feat: add RTK Query fetching to `Search` component
