+++ src/shared/store/store.ts
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit';
-import { SWAPI } from '@shared/api/swapri2';
+import { SWAPI } from '@shared/api';
 import { cardListReducer } from './slices/cardListSlice';
 
 export const store = configureStore({
